Move SearchSerializer from burnplanning to main app.
