+++ govapp/apps/burnplanning/serializers.py
@@ -1,9 +1,7 @@
 import logging
 
-from django.urls import reverse
 from rest_framework import serializers
 
-from govapp.apps.main.serializer import ContentTypeSerializerMixin
 from govapp.apps.main.serializers import (
     DistrictSerializer,
     GenericKeyValueSerializer,
@@ -47,36 +45,6 @@ class BurnPlanElementDatatableSerializer(BurnPlanElementSerializer):
         fields = "__all__"
         # exclude = (
         #     [] #TODO: Add any fields that are not used in the table so query is faster
-
-
-class SearchSerializer(ContentTypeSerializerMixin, serializers.Serializer):
-    reference_number = serializers.CharField(allow_null=True, required=False)
-    name = serializers.CharField(allow_null=True, required=False)
-    status = serializers.CharField(allow_null=True, required=False)
-    status_display = serializers.CharField(
-        source="get_status_display", allow_null=True, required=False
-    )
-    assigned_to = serializers.CharField(allow_null=True, required=False)
-
-    class Meta:
-        model = BurnPlanElement
-        fields = [
-            "id",
-            "reference_number",
-            "name",
-            "status",
-            "status_display",
-            "assigned_to",
-        ]
-
-    def to_representation(self, instance):
-        ret = super().to_representation(instance)
-        verbose_name = instance._meta.verbose_name
-        viewname = verbose_name.replace(" ", "-").lower() + "-detail"
-        ret["link"] = self.context.get("request").build_absolute_uri(
-            reverse(viewname, args=[instance.id])
-        )
-        return ret
 
 
 class BurnPlanUnitSerializer(serializers.ModelSerializer):
